Финальное задание 14 спринта

+++ src/main/java/ru/practicum/shareit/item/storage/InMemoryItemStorage.java
@@ -36,7 +36,7 @@ public class InMemoryItemStorage implements ItemStorage {
     }
 
     @Override
-    public Item getItemById(int id) {
+    public Item getItemById(Long id) {
         return items.get(id);
     }
